tweet_writer/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.4.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.6.15
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.1
cryptography==37.0.1
Flask==1.1.1
frozenlist==1.2.0
google-api-core==2.8.2
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-firestore==2.6.0
googleapis-common-protos==1.56.4
grpcio==1.42.0
grpcio-status==1.41.1
httplib2==0.20.4
idna==2.9
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==5.2.0
oauthlib==3.2.0
packaging==21.3
pip==22.2.2
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
pytest==7.1.2
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.7.2
setuptools==61.2.0
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.25.11
Werkzeug==1.0.1
wheel==0.37.1
yarl==1.6.3
